chore: refine swan package and migration process, begin migrating base package
